Split detector into evidence creation and detection
