+++ light-client/src/misbehavior.rs
@@ -1,2 +1,7 @@
-mod detector;
+mod detect;
 mod error;
+mod evidence;
+
+pub use detect::handle_conflicting_headers;
+pub use error::DetectorError;
+pub use tendermint::evidence::LightClientAttackEvidence;
